Amount使用的类型从byte改为int

+++ Assets/Zephyr/GOAP/Sample/Game/Component/Order/Order.cs
@@ -11,6 +11,6 @@ namespace Zephyr.GOAP.Sample.Game.Component.Order
         public Entity ExecutorEntity;
         public Entity FacilityEntity;
         public FixedString32 ItemName;
-        public byte Amount;
+        public int Amount;
     }
 }
